fix: perf regressions for custom MoEs (#881)
