+++ examples/benchmark/configs/deepseek_v3_te_deepep.yaml
@@ -59,6 +59,7 @@ parallelizer:
   _target_: nemo_automodel.components.moe.parallelizer.parallelize_model
   activation_checkpointing: true
   reshard_after_forward: false
+  wrap_outer_model: false
 
 model:
   _target_: nemo_automodel.components.models.deepseek_v3.model.DeepseekV3ForCausalLM.from_config
