+++ examples/benchmark/configs/gptoss_120b_te_deepep.yaml
@@ -74,7 +74,6 @@ model:
     enable_deepep: true
     fake_balanced_gate: true
     enable_hf_state_dict_adapter: false
-    fsdp_optimization_config: null
 
 checkpoint:
   enabled: false
