+++ examples/benchmark/configs/qwen3_moe_235b_te_deepep.yaml
@@ -59,6 +59,7 @@ parallelizer:
   _target_: nemo_automodel.components.moe.parallelizer.parallelize_model
   activation_checkpointing: true
   reshard_after_forward: false
+  wrap_outer_model: false
 
 model:
   _target_: nemo_automodel.components.models.qwen3_moe.model.Qwen3MoeForCausalLM.from_config
